feat(probe): build for Android and iOS via build-time linking
qtPilot reaches a desktop app by injecting the probe (DYLD_INSERT_LIBRARIES,
LD_PRELOAD, Detours). Neither mobile platform permits inserting a library into a
third-party app, so that route is closed there. The probe already supports a
second delivery mode -- build-time linking, where Q_COREAPP_STARTUP_FUNCTION
starts it once QCoreApplication exists -- and that mode works fine on a device.
This makes it actually buildable for those targets.

Four things stood in the way:

- probe_init_macos.cpp was guarded on Q_OS_MACOS, which is NOT defined for iOS.
  The whole file compiled to nothing, so the startup hook never registered and a
  linked probe sat inert with no way to start. The guard now admits iOS; only the
  DYLD path inside is macOS-specific.

- CMake treats Android as UNIX and iOS as both APPLE and UNIX, so every
  desktop-only branch silently applied to mobile. The launcher -- an injector for
  an already-running process -- was being built for targets where injection does
  not exist. Introduced QTPILOT_MOBILE and excluded it.

- The probe was unconditionally SHARED. Injection needs that; linking does not,
  and a dynamic library inside an app bundle means extra embedding and code
  signing for no benefit. QTPILOT_PROBE_STATIC defaults ON for mobile.

- -Wmissing-include-dirs under -Werror fails every Xcode-generator build: Xcode
  puts DerivedSources paths on the command line before they exist, so the error
  has nothing to do with the code. Kept for generators that populate their own
  include dirs.

Apple also enables -Wshorten-64-to-32 for iOS, which surfaced five genuine
qsizetype -> int narrowings that desktop builds never flagged. All are benign
counts, and all are now explicit rather than implicit. Loop bounds that merely
COMPARE an int against size() are untouched -- those promote and do not narrow.

Verified: probe builds for android_arm64_v8a (static), iOS device arm64 (static),
and macOS (unchanged, still SHARED). Tests and the harness app are disabled
automatically on mobile since neither deploys to a device.

What this enables: linking the probe into a development build of a Qt mobile app
and forwarding its WebSocket -- adb forward on Android, iproxy over USB on iOS --
so a device app can be driven by object identity instead of screen coordinates.

+++ src/probe/api/chrome_mode_api.cpp
@@ -672,7 +672,10 @@ void ChromeModeApi::registerFindMethod() {
 
     // Append to existing ref map (preserves refs from prior find/readPage calls)
     QJsonArray matches;
-    int refCounter = s_refToAccessible.size();
+    // qsizetype -> int: a reference counter bounded by the accessible-element
+  // count. Stated explicitly for the stricter warning set Apple applies to
+  // iOS targets.
+  int refCounter = static_cast<int>(s_refToAccessible.size());
     QString queryLower = query.toLower();
 
     findMatchingNodes(rootIface, queryLower, refCounter, matches, 0, 30);
